+++ Test/PublishPowerShellModule.Tests.ps1
@@ -3,16 +3,23 @@ Set-StrictMode -Version 'Latest'
 & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-WhiskeyTest.ps1' -Resolve)
 
 $testRoot = $null
-$apikey = $null
-$apikeyID = $null
-$repositoryName = $null
-$repositoryUri = $null
-$prerelease = $null
 $context = $null
-$credentials = @{ }
+$credentials = @{}
 $failed = $false
-$publishError = $null
-$registerError = $null
+$publishRoot = $null
+$testModuleName = 'MyModule'
+$repoToUnregister = $null
+
+function Get-TestRepositoryFullPath
+{
+    param(
+        [Parameter(Mandatory)]
+        [String] $Name
+    )
+
+    return Join-Path -Path $testRoot -ChildPath $Name
+
+}
 
 function GivenCredential
 {
@@ -25,21 +32,6 @@ function GivenCredential
     )
 
     $credentials[$WithID] = $Credential
-}
-
-function GivenNoApiKey
-{
-    $script:apikey = $null
-    $script:apikeyID = $null
-}
-
-function GivenPrerelease
-{
-    param(
-        $Prerelease
-    )
-
-    $script:prerelease = $Prerelease
 }
 
 function GivenPublishingFails
@@ -66,63 +58,69 @@ function GivenRepository
 {
     param(
         $Named,
-        $Uri
+        $At
     )
 
-    $script:repositoryName = $Named
-    $script:repositoryUri = $Uri
+    if( -not $At )
+    {
+        $At = Get-TestRepositoryFullPath -Name $Named
+    }
+
+    if( -not ([IO.Path]::IsPathRooted($At)) )
+    {
+        $At = Join-Path -Path $script:testRoot -ChildPath $At
+    }
+
+    New-Item -Path $At -ItemType 'Directory' -Force | Out-Null
+    $script:publishRoot = $At
+
+    Register-PSRepository -Name $Named -PublishLocation $At -SourceLocation $At
+    $script:repoToUnregister = $Named
 }
 
-function Initialize-Test
+function Init
 {
     param(
     )
 
-    $script:apikey = 'fubar:snauf'
-    $script:apikeyID = 'PowerShellExampleCom'
-    $script:repositoryName = $null
-    $script:repositoryUri = $null
-    $script:prerelease = $null
     $script:context = $null
     $script:credentials = @{ }
     $script:failed = $false
-    $script:publishError = $null
-    $script:registerError = $null
-
+    $script:repoToUnregister = $null
+    $script:publishRoot = $null
     $script:testRoot = New-WhiskeyTestRoot
 }
 
-function Invoke-Publish
+function WhenPublishing
 {
     [CmdletBinding()]
     param(
-        [switch]$withoutRegisteredRepo,
+        [String] $ToRepo,
 
-        [String]$ForRepositoryNamed,
+        [String] $RepoAt,
 
-        [String]$RepoAtUri,
+        [String] $ForManifestPath,
 
-        [String]$ForManifestPath,
+        [switch] $WithInvalidPath,
 
-        [switch]$WithNoRepositoryName,
+        [switch] $WithNonExistentPath,
 
-        [switch]$withNoProgetURI,
+        [switch] $WithoutPathParameter,
 
-        [switch]$WithInvalidPath,
+        [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingPlainTextForPassword', '')]
+        [String] $WithCredentialID,
 
-        [switch]$WithNonExistentPath,
+        [switch] $WithNoPrereleasePropertyInManifest,
 
-        [switch]$WithoutPathParameter,
+        [String] $WithPrerelease,
 
-        [String]$WithCredentialID,
-
-        [switch]$WithNoPrereleaseProperty
+        [String] $WithApiKey
     )
     
     $version = '1.2.3'
-    if( $prerelease )
+    if( $WithPrerelease )
     {
-        $version = '1.2.3-{0}' -f $prerelease
+        $version = '1.2.3-{0}' -f $WithPrerelease
     }
 
     $script:context = New-WhiskeyTestContext -ForBuildServer `
@@ -132,43 +130,50 @@ function Invoke-Publish
     
     $TaskParameter = @{ }
 
-    if( $ForRepositoryNamed )
+    if( $ToRepo )
     {
-        $TaskParameter['RepositoryName'] = $ForRepositoryNamed;
+        $TaskParameter['RepositoryName'] = $ToRepo
     }
 
     if( $WithInvalidPath )
     {
-        $TaskParameter.Add( 'Path', 'MyModule.ps1' )
-        New-Item -Path $testRoot -ItemType 'file' -Name 'MyModule.ps1'
+        $TaskParameter.Add( 'Path', "$($testModuleName).ps1" )
+        New-Item -Path $testRoot -ItemType 'file' -Name $TaskParameter['Path']
     }
     elseif( $WithNonExistentPath )
     {
-        $TaskParameter.Add( 'Path', 'MyModule.ps1' )
+        $TaskParameter.Add( 'Path', 'NopeISureDoNotExist.psd1' )
     }
     elseif( -not $WithoutPathParameter )
     {
-        $TaskParameter.Add( 'Path', 'MyModule' )
-        New-Item -Path $testRoot -ItemType 'directory' -Name 'MyModule' 
-        $module = Join-Path -Path $testRoot -ChildPath 'MyModule'
+        $TaskParameter.Add( 'Path', $testModuleName )
+        New-Item -Path $testRoot -ItemType 'directory' -Name $testModuleName
+        $module = Join-Path -Path $testRoot -ChildPath $testModuleName
         if( -not $ForManifestPath )
-        {            
+        {
             $prereleaseProperty = "Prerelease = '';"
-            if( $WithNoPrereleaseProperty )
+            if( $WithNoPrereleasePropertyInManifest )
             {
                 $prereleaseProperty = ''
             }
-            New-Item -Path $module -ItemType 'file' -Name 'MyModule.psd1' -Value @"
+
+            New-item -Path $module -ItemType 'File' -Name "$($testModuleName).psm1"
+            New-Item -Path $module -ItemType 'file' -Name "$($testModuleName).psd1" -Value @"
 @{
+    RootModule = '$($testModuleName).psm1';
+
+    Description = '$($testModuleName)';
+
+    Author = '$([Environment]::UserName)';
+
     # Version number of this module.
-    ModuleVersion = '0.2.0'
+    ModuleVersion = '0.2.0';
 
     PrivateData = @{
         PSData = @{
             $($prereleaseProperty)
         }
     }
-    
 }
 "@
         }
@@ -181,59 +186,12 @@ function Invoke-Publish
     Import-WhiskeyTestModule -Name 'PackageManagement','PowerShellGet'
     Mock -CommandName 'Get-PackageProvider' -ModuleName 'Whiskey'
 
-    $repoName = $script:repositoryName
-    $repoUri = $script:repositoryUri
-
-    Mock -CommandName 'Get-PSRepository' -ModuleName 'Whiskey' -MockWith {
-        return [pscustomobject]@{
-            'Name' = $repoName;
-            'SourceLocation' = $repoUri;
-        }
-    }.GetNewClosure()
-
-    Add-Type -AssemblyName System.Net.Http
-
-    $mock = { }
-    if( $publishError )
-    {
-        $message = $publishError
-        $mock = { 
-                    [CmdletBinding()]
-                    param(
-                        $NuGetApiKey,
-                        $Repository,
-                        $Path,
-                        [switch]$Force
-                    ) 
-                    Write-Error -Message $message
-                }.GetNewClosure()
-    }
-    Mock -CommandName 'Publish-Module' -ModuleName 'Whiskey' -MockWith $mock
-
-    $mock = { }
-    if( $registerError )
-    {
-        $message = $registerError
-        $mock = { 
-                    [CmdletBinding()]
-                    param(
-                        $InstallationPolicy,
-                        $SourceLocation,
-                        $PackageManagementProvider,
-                        $Name,
-                        $PublishLocation
-                    ) 
-                    Write-Error -Message $message
-                }.GetNewClosure()
-    }
-    Mock -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -MockWith $mock
-    
     $Global:Error.Clear()
     $script:failed = $False
 
-    if( $RepoAtUri )
+    if( $RepoAt )
     {
-        $TaskParameter['RepositoryUri'] = $RepoAtUri
+        $TaskParameter['RepositoryLocation'] = $RepoAt
     }
 
     if( $WithCredentialID )
@@ -241,17 +199,23 @@ function Invoke-Publish
         $TaskParameter['CredentialID'] = $WithCredentialID
     }
 
-    if( $apikeyID )
+    if( $WithApiKey )
     {
-        $TaskParameter['ApiKeyID'] = $apikeyID
-        Add-WhiskeyApiKey -Context $context -ID $apikeyID -Value $apikey
+        $TaskParameter['ApiKeyID'] = [Guid]::NewGuid().ToString()
+        Add-WhiskeyApiKey -Context $context -ID $TaskParameter['ApiKeyID'] -Value $WithApiKey
     }
 
     foreach( $key in $credentials.Keys )
     {
         Add-WhiskeyCredential -Context $context -ID $key -Credential $credentials[$key]
     }
-    
+
+    if( $script:repoToUnregister )
+    {
+        Mock -CommandName 'Register-PSRepository' -ModuleName 'Whiskey'
+    }
+
+    Mock -CommandName 'Unregister-PSRepository' -ModuleName 'Whiskey'
     try
     {
         Invoke-WhiskeyTask -TaskContext $context -Parameter $TaskParameter -Name 'PublishPowerShellModule'
@@ -259,14 +223,20 @@ function Invoke-Publish
     catch
     {
         $script:failed = $true
-        Write-Error -ErrorRecord $_
+        Write-Error -ErrorRecord $_ -ErrorAction $ErrorActionPreference
     }
 }
 
 function Reset
 {
     Reset-WhiskeyTestPSModule
+    Get-PSRepository | Where-Object 'Name' -Like 'Whiskey*' | Unregister-PSRepository
+    if( $script:repoToUnregister )
+    {
+        Unregister-PSRepository -Name $script:repoToUnregister
+    }
 }
+Reset
 
 function ThenFailed
 {
@@ -280,127 +250,77 @@ function ThenFailed
 }
 
 function ThenModuleNotPublished
-{    
-    Assert-MockCalled -CommandName 'Publish-Module' -ModuleName 'Whiskey' -Times 0
-}
-
-function ThenRepositoryChecked
 {
-    Assert-MockCalled -CommandName 'Get-PSRepository' -ModuleName 'Whiskey' -Times 1
+    Join-Path -Path $script:publishRoot -ChildPath '*.nupkg' | Should -Not -Exist
 }
 
-function ThenRepositoryNotChecked
-{
-    Assert-MockCalled -CommandName 'Get-PSRepository' -ModuleName 'Whiskey' -Times 0
-}
-
-function ThenRepositoryNotRegistered
-{
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 0
-}
-
-function ThenRepositoryRegistered
+function ThenRepository
 {
     param(
         [Parameter(Mandatory)]
-        [String]$Named,
+        [String] $Named,
 
-        [Parameter(Mandatory)]
-        [String]$AtUri,
+        [switch] $Exists,
 
-        [pscredential]$WithCredential
+        [switch] $NotExists,
+
+        [switch] $NotRegistered,
+
+        [switch] $NotUnregistered
     )
 
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('Repository Name                 expected {0}' -f $Named)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $Name)
-        $Name -eq $Named
-    }
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('Source Location                 expected {0}' -f $AtUri)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $SourceLocation)
-        $AtUri -eq $SourceLocation
-    }
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('Publish Location                expected {0}' -f $AtUri)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $PublishLocation)
-        $AtUri -eq $PublishLocation
-    }
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter { $InstallationPolicy -eq 'Trusted' }
-    Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter { $PackageManagementPRovider -eq 'NuGet' }
-
-    if( $WithCredential )
+    if( $NotRegistered )
     {
-        Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter { $Credential.UserName -eq $WithCredential.UserName }
-        Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 1 -ParameterFilter { $Credential.GetNetworkCredential().Password -eq $WithCredential.GetNetworkCredential().Password }
+        Assert-MockCalled -CommandName 'Register-PSRepository' -ModuleName 'Whiskey' -Times 0
+    }
+
+    if( $Exists )
+    {
+        Get-PSRepository |
+            Where-Object 'Name' -EQ $Named |
+            Where-Object 'PublishLocation' -EQ $script:publishRoot |
+            Should -Not -BeNullOrEmpty
+    }
+
+    if( $NotExists )
+    {
+        Get-PSRepository |
+            Where-Object 'Name' -EQ $Named |
+            Should -BeNullOrEmpty
+    }
+
+    if( $NotUnregistered )
+    {
+        Assert-MockCalled -CommandName 'Unregister-PSRepository' -ModuleName 'Whiskey' -Times 0
     }
 }
 
 function ThenModulePublished
 {
     param(
-        [Parameter(Mandatory)]
-        [String]$ToRepositoryNamed,
-
-        [String]$ExpectedPathName = (Join-Path -Path '.' -ChildPath 'MyModule'),
-
-        [switch]$WithNoRepositoryName
+        $To = $script:context.OutputDirectory.FullName,
+        $WithPrerelease = ''
     )
     
-    Assert-MockCalled -CommandName 'Get-PackageProvider' -ModuleName 'Whiskey' -ParameterFilter { $Name -eq 'NuGet' }
-    Assert-MockCalled -CommandName 'Get-PackageProvider' -ModuleName 'Whiskey' -ParameterFilter { $ForceBootstrap }
-    $expectedApiKey = $apikey
-    Assert-MockCalled -CommandName 'Publish-Module' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('Path Name                       expected {0}' -f $ExpectedPathName)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $Path)
-        
-        $Path -eq $ExpectedPathName
+    if( -not [IO.Path]::IsPathRooted($To) )
+    {
+        $To = Join-Path -Path $script:testRoot -ChildPath $To
     }
-    Assert-MockCalled -CommandName 'Publish-Module' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('Repository Name                 expected {0}' -f $ToRepositoryNamed)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $Repository)
-        $Repository -eq $ToRepositoryNamed
-    }
-    Assert-MockCalled -CommandName 'Publish-Module' -ModuleName 'Whiskey' -Times 1 -ParameterFilter {
-        #$DebugPreference = 'Continue'
-        Write-WhiskeyDebug -Message ('ApiKey                          expected {0}' -f $expectedApiKey)
-        Write-WhiskeyDebug -Message ('                                actual   {0}' -f $NuGetApiKey)
-        $NuGetApiKey -eq $expectedApiKey
-    }
-    Assert-MockCalled -CommandName 'Publish-Module' -ModuleName 'Whiskey' -Times 1 -ParameterFilter { $Force }
+
+    Join-Path -Path $To -ChildPath "$($testModuleName).*.*.*$($WithPrerelease).nupkg" | Should -Exist
 }
 
 function ThenManifest
 {
     param(
-        [String]$ManifestPath = (Join-Path -Path $testRoot -ChildPath 'MyModule\MyModule.psd1'),
-
-        [String]$AtVersion,
+        [String]$ManifestPath = (Join-Path -Path $testRoot -ChildPath "$($testModuleName)\$($testModuleName).psd1"),
 
         [String]$HasPrerelease
     )
 
-    if( -not $AtVersion )
-    {
-        $AtVersion = '{0}.{1}.{2}' -f $context.Version.SemVer2.Major, $context.Version.SemVer2.Minor, $context.Version.SemVer2.Patch
-    }
-
-    $manifest = Test-ModuleManifest -Path $ManifestPath
-
-    $manifest.Version | Should -Be $AtVersion
-    if( $HasPrerelease )
-    {
-        $manifest.PrivateData.PSData.Prerelease | Should -Be $HasPrerelease
-    }
-    else
-    {
-        $manifest.PrivateData.PSData.Prerelease | Should -BeNullOrEmpty
-    }
+    Test-ModuleManifest -Path $ManifestPath | Should -Not -BeNullOrEmpty
+    # Test-ModuleManifest caches manifests if they have a RootModule property.
+    Get-Content -Raw -Path $ManifestPath | Should -Match ([regex]::Escape("Prerelease = '$($HasPrerelease)'"))
 }
 
 function ThenSucceeded
@@ -415,44 +335,41 @@ function Get-Error
     param(
     )
 
+    # PackageManagement and PowerShellGet leave handled errors in Global:Error so we need to filter those errors out.
     $Global:Error |
-        Where-Object {
-            if( -not $_.TargetObject -or -not $_.ScriptStackTrace )
-            {
-                return $true
-            }
-
-            # These errors are internal and can be ignored.
-            $fromGetPackageSource = $_.TargetObject.ToString() -eq 'Microsoft.PowerShell.PackageManagement.Cmdlets.GetPackageSource'
-            $fromResolvingDynamicParams = $_.ScriptStackTrace -match '\bGet-DynamicParameters\b'
-            return -not ( $fromGetPackageSource -and $fromResolvingDynamicParams )
-        }
+        Where-Object 'ScriptStackTrace' -NotMatch '\bPowerShellGet\b' -ErrorAction Ignore |
+        Where-Object 'ScriptStackTrace' -NotMatch '\bPackageManagement\b' -ErrorAction Ignore
 }
 
-Describe 'PublishPowerShellModule.when publishing new module' {
+Describe 'PublishPowerShellModule.when publishing to repository that already exists' {
     AfterEach { Reset }
-    It 'should publish the module' {
-        Initialize-Test
-        GivenRepository 'FubarSnafu'
-        Invoke-Publish -ForRepositoryNamed 'FubarSnafu'
+    It 'should publish the module wihtout registering the repository' {
+        Init
+        GivenRepository 'R1'
+        WhenPublishing -ToRepo 'R1'
         ThenSucceeded
-        ThenRepositoryChecked
-        ThenRepositoryNotRegistered
-        ThenModulePublished -ToRepositoryNamed 'FubarSnafu'
+        ThenModulePublished -To 'R1'
+        ThenRepository 'R1' -Exists -NotRegistered
+    }
+}
+
+Describe 'PublishPowerShellModule.when publishing to repository that does not exist' {
+    AfterEach { Reset }
+    It 'should fail' {
+        Init
+        WhenPublishing -ToRepo 'R2' -ErrorAction Silently
+        ThenFailed 'a repository with that name doesn''t exist'
+        ThenRepository 'R2' -NotExists -NotUnregistered
     }
 }
 
 Describe 'PublishPowerShellModule.when publishing prerelease module' {
     AfterEach { Reset }
     It 'should succeed' {
-        Initialize-Test
-        GivenRepository 'SomeRepo'
-        GivenPrerelease 'beta1'
-        Invoke-Publish -ForRepositoryNamed 'SomeRepo'
+        Init
+        WhenPublishing -WithPrerelease 'beta1'
         ThenSucceeded
-        ThenRepositoryChecked
-        ThenRepositoryNotRegistered
-        ThenModulePublished -ToRepositoryNamed 'SomeRepo'
+        ThenModulePublished -WithPrerelease '-beta1'
         ThenManifest -HasPrerelease 'beta1'
     }
 }
@@ -460,170 +377,82 @@ Describe 'PublishPowerShellModule.when publishing prerelease module' {
 Describe 'PublishPowerShellModule.when publishing prerelease module but module manifest is missing Prerelease element' {
     AfterEach { Reset }
     It 'should fail' {
-        Initialize-Test
-        GivenPrerelease 'rc5'
-        GivenRepository 'SomeRepo'
-        Invoke-Publish -ForRepositoryNamed 'SomeRepo' -WithNoPrereleaseProperty -ErrorAction SilentlyContinue
+        Init
+        WhenPublishing -WithPrerelease 'rc5' -WithNoPrereleasePropertyInManifest -ErrorAction SilentlyContinue
         ThenFailed 'missing a "Prerelease" property'
-        ThenRepositoryNotChecked
-        ThenRepositoryNotRegistered
-        ThenModuleNotPublished
     }
 }
 
-Describe 'PublishPowerShellModule.when publishing with no repository name' {
+Describe 'PublishPowerShellModule.when publishing with no repository name or repository location' {
     AfterEach { Reset }
-    It 'should fail' {
-        Initialize-Test
-        Invoke-Publish -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'Property\ "RepositoryName"\ is mandatory'
-        ThenRepositoryNotChecked
-        ThenRepositoryNotRegistered
-        ThenModuleNotPublished
-    }
-}
-
-Describe 'PublishPowerShellModule.when publishing fails' {
-    AfterEach { Reset }
-    It 'should fail' {
-        Initialize-Test
-        GivenRepository 'FubarSnafu'
-        GivenPublishingFails -WithError 'something went wrong!'
-        Invoke-Publish -ForRepositoryNamed 'FubarSnafu' -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'something\ went\ wrong\!'
-    }
-}
-
-Describe 'PublishPowerShellModule.when publishing to a new repository' {
-    AfterEach { Reset }
-    It 'should succeed' {
-        Initialize-Test
-        Invoke-Publish -ForRepositoryName 'ANewRepo' -RepoAtUri 'https://example.com' 
-        ThenRepositoryChecked
-        ThenRepositoryRegistered 'ANewRepo' -AtUri 'https://example.com/'
-        ThenModulePublished -ToRepositoryNamed 'ANewRepo'
+    It 'should publish to build output directory' {
+        Init
+        WhenPublishing
         ThenSucceeded
+        ThenRepository 'Whiskey*' -NotExists
+        ThenModulePublished -To $context.OutputDirectory
     }
 }
 
-Describe 'PublishPowerShellModule.when publishing to a new repository that requires a credential' {
+Describe 'PublishPowerShellModule.when given an API key' {
     AfterEach { Reset }
-    It 'should succeed' {
-        Initialize-Test
-        $cred = New-Object 'pscredential' ('fubar',(ConvertTo-SecureString -String 'snafu' -AsPlainText -Force))
-        GivenCredential $cred -WithID 'somecred'
-        Invoke-Publish -ForRepositoryName 'ANewRepo' -RepoAtUri 'https://example.com' -WithCredentialID 'somecred'
+    It 'should fail' {
+        Init
+        Mock -CommandName 'Publish-Module' `
+             -ModuleName 'Whiskey' `
+             -ParameterFilter { $NuGetApiKey -eq 'API6' } `
+             -Verifiable
+        WhenPublishing -WithApiKey 'API6'
         ThenSucceeded
-        ThenRepositoryChecked
-        ThenRepositoryRegistered 'ANewRepo' -AtUri 'https://example.com/' -WithCredential $cred
-        ThenModulePublished -ToRepositoryNamed 'ANewRepo'
-    }
-}
-
-Describe 'PublishPowerShellModule.when publishing to a new repository but its URI is not given' {
-    AfterEach { Reset }
-    It 'should fail' {
-        Initialize-Test 
-        Invoke-Publish -ForRepositoryNamed 'Fubar' -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'Property\ "RepositoryUri"\ is\ mandatory'
-        ThenRepositoryChecked
-        ThenRepositoryNotRegistered
-        ThenModuleNotPublished
-    }
-}
-
-Describe 'PublishPowerShellModule.when registering a repository fails' {
-    AfterEach { Reset }
-    It 'should fail' {
-        Initialize-Test
-        GivenRegisteringFails -WithError 'something went wrong!'
-        Invoke-Publish -ForRepositoryNamed 'FubarSnafu' -RepoAtUri 'https://example.com' -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'something\ went\ wrong\!'
-    }
-}
-
-Describe 'PublishPowerShellModule.when no API key' {
-    AfterEach { Reset }
-    It 'should fail' {
-        Initialize-Test
-        GivenNoApiKey
-        GivenRepository 'Fubar'
-        Invoke-Publish -ForRepositoryNamed 'Fubar'  -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'Property\ "ApiKeyID"\ is\ mandatory'
-        ThenRepositoryNotChecked
-        ThenRepositoryNotRegistered
-        ThenModuleNotPublished
+        Assert-VerifiableMock
     }
 }
 
 Describe 'PublishPowerShellModule.when path parameter is not included' {
     AfterEach { Reset }
     It 'should fail' {
-        Initialize-Test
-        GivenRepository 'Fubar'
-        Invoke-Publish -ForRepositoryNamed 'Fubar' -WithoutPathParameter -ErrorAction SilentlyContinue
+        Init
+        WhenPublishing -WithoutPathParameter -ErrorAction SilentlyContinue
         ThenFailed -WithError '"Path\b.*\bis mandatory'
-        ThenRepositoryNotChecked
-        ThenRepositoryNotRegistered
-        ThenModuleNotPublished
     }
 }
 
-Describe 'PublishPowerShellModule.when non-existent path parameter' {
+Describe 'PublishPowerShellModule.when path does not exist' {
     AfterEach { Reset }
     It 'should fail' {
-        Initialize-Test
-        GivenRepository 'Fubar'
-        Invoke-Publish -ForRepositoryNamed 'Fubar' -WithNonExistentPath -ErrorAction SilentlyContinue
+        Init
+        WhenPublishing -WithNonExistentPath -ErrorAction SilentlyContinue
         ThenFailed -WithError 'does\ not\ exist'
-        ThenModuleNotPublished
     }
 }
 
-Describe 'PublishPowerShellModule.when non-directory path parameter' {
+Describe 'PublishPowerShellModule.when path is not a directory' {
     AfterEach { Reset }
     It 'should fail' {
-        Initialize-Test
-        GivenRepository 'Fubar'
-        Invoke-Publish -ForRepositoryNamed 'Fubar' -WithInvalidPath  -ErrorAction SilentlyContinue
+        Init
+        WhenPublishing -WithInvalidPath  -ErrorAction SilentlyContinue
         ThenFailed -WithError 'should resolve to a directory'
-        ThenModuleNotPublished
     }
 }
 
-Describe 'PublishPowerShellModule.when invalid manifest' {
+Describe 'PublishPowerShellModule.when module manifest is invalid' {
     AfterEach { Reset }
     It 'should fail' {
-        Initialize-Test
-        GivenRepository 'Fubar'
-        Invoke-Publish -ForRepositoryNamed 'Fubar' -withoutRegisteredRepo -ForManifestPath 'fubar' -ErrorAction SilentlyContinue
-        ThenFailed -WithError 'path\ "fubar"\ either\ does\ not\ exist'
-        ThenModuleNotPublished
+        Init
+        WhenPublishing -ForManifestPath 'fubar' -ErrorAction SilentlyContinue
+        ThenFailed -WithError '"fubar"\ does\ not\ exist'
     }
 }
 
-Describe 'PublishPowerShellModule.when registering an existing PSRepository under a different name' {
+Describe 'PublishPowerShellModule.when given repository by location and by name' {
     AfterEach { Reset }
-    It 'should use already registered PSRepository' {
-        Initialize-Test
-        GivenRepository -Named 'FirstRepo' -Uri 'https://example.com'
-        Invoke-Publish -ForRepositoryNamed 'ImposterRepo' -RepoAtUri 'https://example.com'
+    It 'should use repository registered by location' {
+        Init
+        GivenRepository -Named 'R7' -At 'R7'
+        WhenPublishing -ToRepo 'R8' -RepoAt $script:publishRoot
         ThenSucceeded
-        ThenRepositoryNotRegistered
-        ThenRepositoryChecked
-        ThenModulePublished -ToRepositoryNamed 'FirstRepo'
-    }
-}
-
-Describe 'PublishPowerShellModule.when re-registering an existing PSRepository under the same name' {
-    AfterEach { Reset }
-    It 'should succeed' {
-        Initialize-Test
-        GivenRepository -Named 'FirstRepo' -Uri 'https://example.com'
-        Invoke-Publish -ForRepositoryNamed 'FirstRepo' -RepoAtUri 'https://example.com'
-        ThenSucceeded
-        ThenRepositoryChecked
-        ThenRepositoryNotRegistered
-        ThenModulePublished -ToRepositoryNamed 'FirstRepo'
+        ThenRepository 'R8' -NotRegistered -NotExists
+        ThenRepository 'R7' -NotRegistered -Exists
+        ThenModulePublished -To 'R7'
     }
 }