+++ Test/Version.Tests.ps1
@@ -1,4 +1,7 @@
 
+#Requires -Version 5.1
+Set-StrictMode -Version 'Latest'
+
 & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-WhiskeyTest.ps1' -Resolve)
 
 [Whiskey.Context]$context = $null